+++ src/pages/main/exchange.html
@@ -5,10 +5,15 @@
     <link rel="icon" type="image/svg+xml" href="/vite.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>exchange page</title>
-    <link rel="stylesheet" href="./src/index.css" />
-    <script type="module" src="/src/pages/main/index.js"></script>
+    <link rel="stylesheet" href="/src/style/reset.css" />
+    <link rel="stylesheet" href="/src/style/variable.css" />
+    <!-- <link rel="stylesheet" href="/src/components/Main/State/State.css" /> -->
+    <script type="module" src="/src/pages/main/exchange.js"></script>
   </head>
   <body>
+    <list-item></list-item>
+    <h1>exchange page!</h1>
+
     <main-state status="reserved" price="800000"></main-state>
     <main-state status="complete" price="50000"></main-state>
     <main-state status="" price="150000"></main-state>
