+++ src/pages/main/exchange.js
@@ -1,0 +1,9 @@
+console.log('테스트');
+
+import '/src/components/Main/PlusButton/PlusButton';
+
+import '/src/components/Main/State/State';
+
+import '/src/components/Main/ListItem/ListItem.js';
+
+import '/src/components/Main/ListItem/ListItem.css';
