+++ src/pages/main/index.js
@@ -1,2 +1,1 @@
-import '/src/components/Main/PlusButton';
-import '/src/components/Main/State';
+console.log('test');
